Update play.js
This adds all the code for the statue puzzle, the player gets the key to open the room with the yellow jewel, then finds the blue jewel. if the player possesses both jewels and stands in front of the statue, the statue gets destroyed and reveals the key needed to open the last door.